Agent-base Phase 1: storage split — finish the barrel carve-out, break the repository ⇄ docsIngest cycle (item 4)
- `pageKeyOf` moves verbatim out of docsIngest.ts into the new import-free
  leaf src/context/docTitles.ts; docsIngest.ts and the storage layer both
  import it from there, so neither imports the other.
- src/context/* now imports the owning repository/ domain modules directly
  instead of the barrel: zero imports from context/ into the barrel, zero
  from the barrel into context/ (the old cycle is gone).
- New src/storage/repository/interactions.ts: the raw interaction archive
  (recordInteraction, searchMemory, recap/tail reads, delete/edit honouring)
  moved verbatim out of repository.ts.
- wasMemberDigestSentRecently/recordMemberDigestSent move verbatim to
  digestAlerts.ts (its "deliberately NOT here" cycle apology is now moot);
  listCuratedKnowledgeCreatedSince/listReleaseWatchUpdatesSince move verbatim
  to knowledge.ts (pageKeyOf now from the docTitles leaf).
- repository.ts is now a PURE `export *` barrel — every import site keeps
  compiling unchanged.
- tests/backgroundJobCost.test.ts additionally mocks the adminStats domain
  module (the spy's target moved with the context repoint; assertions
  unchanged — the established scanner-repoint precedent).
- module-map entries for docTitles.ts + interactions.ts, updated
  repository.ts/repository/ wording, recipes.md barrel row updated.

All moved blocks verified byte-identical against HEAD (diff over the moved
line ranges). typecheck, test, test:security, context:check, build all green.
