+++ docs/agents/module-map.md
@@ -78,6 +78,7 @@ is marked **🔒**. Changes there need a `SECURITY:` test (see
 - `src/config/` — Per-domain zod slice fragments (each var's chain + doc comment lives with its domain) and their slice-local refinements; `env.ts` owns the one dotenv load + blank-normalisation. Adding a setting starts in the right slice here.
 - `src/config/boot.ts` — Boot-path config: validates ONLY the db+log slices so `logger.ts`/`storage/db.ts`/`storage/migrate.ts` run with just `DATABASE_URL` — what lets a bare `npm run migrate` work without the app's other required vars.
 - `src/context/` — The community-context learning loop: nightly digest builder, knowledge refresh, docs ingest, link-rot check, and the PII-scrubbed export that is the DB-to-repo privacy boundary.
+- `src/context/docTitles.ts` — Import-free leaf for docs-ingest chunk-title helpers (`pageKeyOf`); shared by docsIngest.ts and repository/knowledge.ts so neither imports the other (the old repository ⇄ docsIngest cycle).
 - `src/crashHandlers.ts` — Installs handlers for unhandled rejections and uncaught exceptions, so a process death always leaves a logged reason.
 - `src/dailyBudgetNotice.ts` — Static text for the daily reply-budget notice; the 24h debounce window itself is tracked inline by the router.
 - `src/dailyReplyBudgetWarning.ts` — The short warning line appended to a real reply as a member approaches their daily budget, so the cutoff is not the first sign a limit exists.
@@ -105,8 +106,9 @@ is marked **🔒**. Changes there need a `SECURITY:` test (see
 - `src/router.ts` — 🔒 The hot path: every inbound message lands here. Rate limits, budgets, tier resolution, confirm handling, moderation and the agent call are all sequenced in this file.
 - `src/status/` — Anthropic status-page check behind the "is it me or is Anthropic down?" answer, with its own cache so a common question costs nothing.
 - `src/storage/` — 🔒 Postgres + pgvector: the pool, the schema fragments + concatenating migrator, local embeddings, runtime policies, and the repository that owns every query. Admin-facing reads are conversation-scoped in SQL here.
-- `src/storage/repository.ts` — 🔒 The repository entry point every caller imports: still holds the not-yet-extracted queries, and re-exports the per-domain modules in `repository/`. Conversation scoping for admin reads is enforced in the queries themselves, not by callers.
-- `src/storage/repository/` — 🔒 Per-domain query modules being carved out of `repository.ts` one domain at a time (audit L14). Add a new query to its domain module here, not to `repository.ts`; everything is re-exported so import sites never change.
+- `src/storage/repository.ts` — 🔒 The repository barrel every caller imports: PURE `export *` lines over the per-domain modules in `repository/` — no query bodies live here anymore. Conversation scoping for admin reads is enforced in the queries themselves, not by callers.
+- `src/storage/repository/` — 🔒 The per-domain query modules the audit-L14 carve-out produced (complete — `repository.ts` is now only the barrel). Add a new query to its domain module here; everything is re-exported so import sites never change.
+- `src/storage/repository/interactions.ts` — 🔒 The raw interaction archive: recordInteraction, semantic memory search, recap/tail reads, and the platform delete/edit honouring paths (scoped to platform+conversation+message id).
 - `src/storage/repository/projects.ts` — 🔒 Project shared memory (issue #927). `visibleProjectIds` is the one place the two access checks live — membership (expanded through linked `persons`) and surface (a bound conversation or a DM) — and every read/write here goes through it in SQL.
 - `src/storage/repository/whatsappLidMap.ts` — 🔒 Durable WhatsApp LID -> phone mapping. A LID is a privacy id that looks like a number but matches no one; persisting what the adapter learns from real envelopes lets a LID be resolved rather than refused. PII — erased by `forget_me`/`purge_user_data`. See docs/SECURITY.md §6b.
 - `src/storage/schema/` — 🔒 The schema as ordered SQL fragments (00–27 base, 50–53 community, 70 adapter — the gaps are deliberate) concatenated by `manifest.ts` and replayed by `migrate.ts` as ONE atomic query on every deploy. Every statement stays `IF NOT EXISTS`; new fragments must be added to the manifest's explicit array (never a glob — order is load-bearing), and a fragment missing from it fails the sync test.
